:sparkles: feat: Assossiação do modelo do veiculo ao banco com String.

+++ src/main/java/site/SpringWeb/repositorio/ClientesRepo.java
@@ -14,7 +14,7 @@ public interface ClientesRepo extends CrudRepository<Cliente, Integer> {
   @Query(value = "select CASE WHEN count(1) > 0 THEN 'true' ELSE 'false' END from clientes where id = :id", nativeQuery = true)
   public boolean exist(int id);
 
-  Optional<Cliente> findById(int id); // Adicionado para buscar por ID
+  Optional<Cliente> findById(Long id); // Adicionado para buscar por ID
 
   // @Query(value="select * from clientes where email = :email and senha =
   // :senha", nativeQuery = true)
